🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics ✨ feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🐛 fix(local_storage.dart): remove unused player_id field in LocalDatabase table
The player_id field in the LocalDatabase table was not being used and has been removed to improve the database schema.

🐛 fix(local_storage.dart): add error handling for updateData method
Added error handling to the updateData method in the LocalStorage class to catch and print any errors that occur during the update operation.

🐛 fix(party.dart): fix incorrect statusText values in getStatusFromText method
Fixed the incorrect statusText values in the getStatusFromText method of the Party class. The correct values are 'notStarted', 'inProgress', 'won', 'lost', and 'paused'.

🐛 fix(player.dart): remove empty line in Player class
Removed an empty line in the Player class to improve code readability.

✨ feat(player.dart): add fromJson and toJson methods to Player class
Added fromJson and toJson methods to the Player class to enable conversion between JSON and Player objects.

🐛 fix(stats.dart): fix incorrect field name in toJson method
Fixed the incorrect field name in the toJson method of the Stats class. The correct field name is 'score' instead of 'total_score'.

🐛 fix(main.dart): add initialRoute parameter to GetMaterialApp
Added the initialRoute parameter to the GetMaterialApp widget in the main.dart file to specify the initial route of the application.

✨ feat(auth_controller.dart): add continueAsGuest method
Added the continueAsGuest method to the AuthController class to allow users to continue as a guest without creating an account.

✨ feat(auth_controller.dart): add getCurrentPlayer method
Added the getCurrentPlayer method to the AuthController class to retrieve the current player from the local storage.

🐛 fix(evolution_controller.dart): clear regions list before fetching saved stats
Added a line to clear the regions list before fetching the saved stats in the onInit method of the EvolutionController class.

🐛 fix(evolution