🔧 fix(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics ✨ feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🔧 fix(local_storage.dart): change database name to 'database.db' for better clarity
The database name has been changed to 'database.db' to provide a more descriptive and meaningful name.

🔧 fix(local_storage.dart): uncomment deleteDatabase call to ensure database is deleted before initialization
The deleteDatabase call has been uncommented to ensure that the database is deleted before initialization. This helps to avoid any conflicts or issues that may arise from an existing database.

🔧 fix(local_database.dart): add player_id column to stats table for proper data association
The player_id column has been added to the stats table to establish a proper association between players and their respective stats. This allows for more accurate and meaningful data representation.

🔧 fix(local_storage.dart): add methods to check if user is logged in and to set user login status
The isUserLoggedIn method has been added to check if the user is logged in. It returns true if the 'isUserLoggedIn' key exists in the storage and its value is true, otherwise it returns false. The setUserLoggedIn method has been added to set the user login status by writing the 'isUserLoggedIn' key to the storage with the provided value.

🔧 fix(stats.dart): remove trailing comma to fix syntax error
The trailing comma after the 'level' field has been removed to fix the syntax error.

🔧 fix(main.dart): import correct app_routes.dart file
The import statement for the app_routes.dart file has been updated to import the correct file.

🔧 fix(auth_controller.dart): remove unnecessary empty line
The unnecessary empty line has been removed for better code readability.

🔧 fix(evolution_controller.dart): uncomment getCurrentPlayer call to fetch current player
The getCurrentPlayer call has been uncommented to fetch the current player. This ensures that the player's stats are properly updated.

🔧 fix(region_controller.dart): uncomment generateRegions and addParties calls for proper initialization
The generateRegions and addPart