🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics ✨ feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🐛 fix(player.dart): add totalScore field to Player.fromJson constructor
The Player.fromJson constructor now includes the totalScore field to properly parse the total_score value from the JSON response. This ensures that the totalScore field is populated correctly when creating a Player object from JSON data.

✨ feat(evolution_controller.dart): add getRanking method to fetch player rankings
The getRanking method is added to the EvolutionController class to fetch the player rankings from the database. This method queries the "players" table and retrieves the id, name, gender, and total_score fields for each player. The retrieved data is then used to create Player objects and add them to the players list. This allows the leaderboard to display the current rankings of the players.

🐛 fix(region_controller.dart): uncomment generateRegions and addParties methods
The generateRegions and addParties methods are uncommented in the RegionController class. These methods are responsible for generating the initial regions and parties data and inserting them into the local storage. By uncommenting these methods, the regions and parties data will be properly initialized and available for use in the application.

🐛 fix(home_screen.dart): uncomment RegionPreviewWidget
The RegionPreviewWidget is uncommented in the HomeScreen class. This widget displays a preview of the first region in the regions list. By uncommenting this widget, the preview will be visible on the home screen.

🐛 fix(leader_board_page.dart): update Player class import statement
The import statement for the Player class is updated to reflect the correct file path. This ensures that the Player class is properly imported and used in the LeaderboardPage class.

✨ feat(leader_board_page.dart): fetch player rankings and display them in the leaderboard
The LeaderboardPage class now fetches the player rankings from the EvolutionController and displays them in the leaderboard. The rankings are retrieved using the getRanking method and stored in the players list. The leaderboard is then built using the players list,